==> 3.0_Remove_duplicates.ipynb <==
import pandas as pd
import pickle

df_comp_nona = pd.read_pickle("./OUT_dfs/df_comp_full_processed_nona.pkl")
df_appr_nona = pd.read_pickle("./OUT_dfs/df_appr_full_processed_nona.pkl")
df_a = df_appr_nona.copy()
df_c = df_comp_nona.copy()

# Create unique comp ID

comp_columns_to_look_for_duplicates_in = ['ADDRESS1', 'CITY', 'STATE', 'ZIPCODE',"SALEDATE"]
new_id = df_c.groupby(comp_columns_to_look_for_duplicates_in).ngroup()
df_c.insert(2, 'UNIQUECOMPIDNEW', new_id)

uniqueRows = df_c.drop_duplicates(subset=["UNIQUECOMPIDNEW"])

# Create a mapping dictionary

dict_apprid_to_uniquecompidnew = df_c.groupby('SUBJ_APPR_ID')['UNIQUECOMPIDNEW'].apply(list).to_dict()


# save pickle dictionary file
pickle.dump(dict_apprid_to_uniquecompidnew, open("./OUT_dfs/dict_apprid_to_uniquecompidnew", "wb"))
# to load :
#dict_apprid_to_uniquecompidnew = pickle.load(open("./OUT_dfs/dict_apprid_to_uniquecompidnew", "wb"))

dict_apprid_to_uniquecompidnew[857100]

# Remove duplicate rows from  from Comp dataframe to get a proper X dataframe for unsupervised learning

uniqueComps = df_c.drop_duplicates(subset=["UNIQUECOMPIDNEW"])

uniqueComps.to_pickle("./OUT_dfs/uniqueComps.pkl")
